Add shutdown logic to container
When shutdown is called on a container it will search through all
instances for a function with the name `__amendShutdown` and call it.
This is not an optimal solution since it will leave some footprint on
modules using this feature. The footprint is rathe small thoug and
should not cause too much trouble. But if a better solution arises it
should be replaced.