Rewrite of fromConfig
The way to instantiate containers have been completely rewritten from
scratch. All parts are separated into different modules namely:

 * makePath
 * evaluateType
 * getPath
 * findModule
 * populateDI
 * addParent

The paths to the different modules are now evaluated more consistently
and the entire project is more stable. Parent support has been moved to
the config files and will be removed from the container in the future.